:bug: fix(calendar-event-transition): update tests
