+++ packages/core/src/__mocks__/v1/events/events.misc.ts
@@ -1,3 +1,4 @@
+import { ObjectId } from "bson";
 import { Origin, Priorities } from "@core/constants/core.constants";
 import { Schema_Event } from "@core/types/event.types";
 
@@ -5,10 +6,10 @@ import { Schema_Event } from "@core/types/event.types";
  * Assortment of events with no direct relation to another
  */
 
-const USER = "6227e1a1623abad10d70afbf";
+const USER = new ObjectId().toString();
 
 export const CHILL_ALL_DAY: Schema_Event = {
-  _id: "620c177bfadfdec705cdd69c",
+  _id: new ObjectId().toString(),
   gEventId: "6csjad336cs3ibb469imcb9kc9gj6bb26ss30bb56kojgoj464o66ohh60",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
@@ -20,7 +21,7 @@ export const CHILL_ALL_DAY: Schema_Event = {
   endDate: "2022-09-24",
 };
 export const CLIMB: Schema_Event = {
-  _id: "62322b127837957382660217",
+  _id: new ObjectId().toString(),
   gEventId: "ccq34eb261j3ab9jckpj6b9kcos6cbb26pi38b9pc5i64e9mcgp3ao9p6o",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
@@ -32,7 +33,7 @@ export const CLIMB: Schema_Event = {
   endDate: "2022-03-01T19:00:00-06:00",
 };
 export const EUROPE_TRIP: Schema_Event = {
-  _id: "6262d2840138892cb743444a",
+  _id: new ObjectId().toString(),
   user: USER,
   origin: Origin.COMPASS,
   title: "Europe Trip",
@@ -44,7 +45,7 @@ export const EUROPE_TRIP: Schema_Event = {
   order: 2,
 };
 export const LEARN_CHINESE: Schema_Event = {
-  _id: "awk92akknm",
+  _id: new ObjectId().toString(),
   description: "",
   isSomeday: true,
   origin: Origin.COMPASS,
@@ -56,7 +57,7 @@ export const LEARN_CHINESE: Schema_Event = {
   order: 1,
 };
 export const MARCH_1: Schema_Event = {
-  _id: "62322b127837957382660212",
+  _id: new ObjectId().toString(),
   gEventId: "2ip0l4k0kqhg22cagtmrlml5mn",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
@@ -68,7 +69,7 @@ export const MARCH_1: Schema_Event = {
   endDate: "2022-03-02",
 };
 export const GROCERIES: Schema_Event = {
-  _id: "620c177bfadfdec705cdd70a",
+  _id: new ObjectId().toString(),
   gEventId: "pihjll1k75s1g9019ru6tkb97c",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
@@ -80,7 +81,7 @@ export const GROCERIES: Schema_Event = {
   endDate: "2022-02-21T12:45:00-06:00",
 };
 export const MULTI_WEEK: Schema_Event = {
-  _id: "62322b12783795738266022e",
+  _id: new ObjectId().toString(),
   gEventId: "1k2rgneltn79cbchccut08alqs",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
@@ -93,7 +94,7 @@ export const MULTI_WEEK: Schema_Event = {
   endDate: "2022-09-22T00:00:00-06:00",
 };
 export const TY_TIM: Schema_Event = {
-  _id: "62322b127837957382660231",
+  _id: new ObjectId().toString(),
   gEventId: "726v6dgnasekgmv5hc1jifpumm",
   user: USER,
   origin: Origin.GOOGLE_IMPORT,
