+++ packages/core/src/__tests__/core.test.start.ts
@@ -1,0 +1,3 @@
+import { mockBSON } from "@core/__tests__/mock.setup";
+
+mockBSON();
