+++ packages/web/src/__tests__/web.test.init.js
@@ -1,2 +1,6 @@
+import { mockBSON } from "@core/__tests__/mock.setup";
+
 process.env.API_BASEURL = "http://localhost:3000/api";
 process.env.GOOGLE_CLIENT_ID = "mockedClientId";
+
+mockBSON();
