+++ packages/web/src/__tests__/web.test.start.js
@@ -1,12 +1,9 @@
 import { server } from "@web/__tests__/__mocks__/server/mock.server";
 import {
   clearLocalStorageMock,
-  mockBSON,
   mockResizeObserver,
   mockScroll,
 } from "@web/__tests__/utils/test.util";
-
-mockBSON();
 
 beforeAll(() => {
   mockScroll();
